[Fix #159402806] Add packages to handle upload on both client and server
