+++ client/client.js
@@ -5,10 +5,13 @@ import {
   ApolloLink,
 } from 'apollo-boost';
 
+import { createUploadLink } from 'apollo-upload-client';
+
 const cache = new InMemoryCache();
 
 const httpLink = new HttpLink({ uri: process.env.SERVER_API_BASE_URL });
 
+const uploadLink = createUploadLink({ uri: process.env.SERVER_API_BASE_URL });
 
 const middlewareAuthLink = new ApolloLink((operation, forward) => {
   // The token should be gotten from the redux store or which ever location it is stored at login
@@ -17,8 +20,7 @@ const middlewareAuthLink = new ApolloLink((operation, forward) => {
   return forward(operation);
 });
 
-const httpLinkWithAuthToken = middlewareAuthLink.concat(httpLink);
-
+const httpLinkWithAuthToken = middlewareAuthLink.concat(uploadLink);
 
 const Client = new ApolloClient({
   link: httpLinkWithAuthToken,
