+++ client/package.json
@@ -20,6 +20,7 @@
   "license": "MIT",
   "dependencies": {
     "apollo-boost": "^0.1.10",
+    "apollo-upload-client": "^8.1.0",
     "axios": "0.16.2",
     "babel-core": "6.26.0",
     "babel-loader": "7.1.2",
