+++ server/api/views.py
@@ -26,8 +26,6 @@ from .serializers import CategorySerializer, EventSerializer,\
 from .models import Category, Interest, Event, Attend, AndelaUserProfile
 from .utils.oauth_helper import save_credentials
 
-from .utils.upload_helper import file_uploader
-
 from .setpagination import LimitOffsetpage
 from .slack import get_slack_id, notify_channel, notify_user
 
@@ -209,8 +207,6 @@ class CreateEventView(CreateAPIView):
             social_event = Category.objects.get(id=int(social_event_id))
         except Category.DoesNotExist:
             raise Http404\
-
-        # file_uploader
 
         new_event = Event(
             title=body_data.get('title'),
