+++ server/requirements.txt
@@ -1,25 +1,36 @@
-Django==1.11.5
-PyJWT==1.5.3
-Rx==1.6.1
 aniso8601==3.0.2
+appdirs==1.4.3
 asn1crypto==0.24.0
+astroid==2.0.4
+attrs==18.2.0
+black==18.6b4
+cachetools==2.1.0
 certifi==2018.4.16
 cffi==1.11.5
 chardet==3.0.4
+click==6.7
 coverage==4.5.1
 cryptography==2.2.2
 dj-database-url==0.4.2
+Django==1.11.5
 django-cors-headers==2.1.0
 django-crontab==0.7.1
 django-filter==1.0.4
 django-webpack-loader==0.5.0
-djangorestframework-jwt==1.11.0
 djangorestframework==3.6.3
+djangorestframework-jwt==1.11.0
 flake8==3.5.0
+google-api-core==1.4.0
 google-api-python-client==1.6.2
+google-auth==1.5.1
 google-auth-oauthlib==0.2.0
-graphene-django==2.1rc1
+google-cloud-core==0.28.1
+google-cloud-storage==1.11.0
+google-resumable-media==0.3.1
+googleapis-common-protos==1.5.3
 graphene==2.1.2
+graphene-django==2.1rc1
+graphene-file-upload==1.0.0
 graphql-core==2.1rc3
 graphql-relay==0.4.5
 gunicorn==19.7.1
@@ -27,33 +38,45 @@ hashids==1.2.0
 httplib2==0.11.3
 idna==2.6
 iso8601==0.1.12
+isort==4.3.4
 jsonpickle==0.9.6
+lazy-object-proxy==1.3.1
 mccabe==0.6.1
 mock==2.0.0
 oauth2client==4.1.2
+oauthlib==2.1.0
+pbr==4.2.0
 pluggy==0.6.0
 promise==2.1
+protobuf==3.6.1
 psycopg2==2.7.3.1
 py==1.5.4
-pyasn1-modules==0.2.1
 pyasn1==0.4.3
+pyasn1-modules==0.2.1
 pycodestyle==2.3.1
 pycparser==2.18
 pyflakes==1.6.0
+PyJWT==1.5.3
+pylint==2.1.1
 python-dateutil==2.7.3
 python-env==1.0.0
 pytz==2018.4
 requests==2.18.4
+requests-oauthlib==1.0.0
 rsa==3.4.2
+Rx==1.6.1
 singledispatch==3.4.0.3
 six==1.11.0
 slackclient==1.0.9
 snapshottest==0.5.0
 termcolor==1.1.0
+toml==0.9.6
 tox==3.0.0
+typed-ast==1.1.0
 typing==3.6.4
 uritemplate==3.0.0
 urllib3==1.22
 virtualenv==16.0.0
 websocket-client==0.44.0
 whitenoise==3.3.0
+wrapt==1.10.11
